+++ src/caselawclient/Client.py
@@ -17,6 +17,7 @@ from requests.structures import CaseInsensitiveDict
 from requests_toolbelt.multipart import decoder
 
 from caselawclient import xquery_type_dicts as query_dicts
+from caselawclient.client_helpers import VersionAnnotation
 from caselawclient.models.documents import (
     DOCUMENT_COLLECTION_URI_JUDGMENT,
     DOCUMENT_COLLECTION_URI_PRESS_SUMMARY,
@@ -458,50 +459,83 @@ class MarklogicApiClient:
         self,
         judgment_uri: DocumentURIString,
         judgment_xml: bytes,
-        annotation: Optional[str] = None,
+        annotation: VersionAnnotation,
     ) -> requests.Response:
         """assumes the judgment is already locked, does not unlock/check in
         note this version assumes the XML is raw bytes, rather than a tree..."""
+
         validate_content_hash(judgment_xml)
         uri = self._format_uri_for_marklogic(judgment_uri)
+
+        annotation.set_calling_function("save_locked_judgment_xml")
+
         vars: query_dicts.UpdateLockedJudgmentDict = {
             "uri": uri,
             "judgment": judgment_xml.decode("utf-8"),
-            "annotation": annotation or "edited by save_locked_judgment_xml",
+            "annotation": annotation.as_json,
         }
 
         return self._send_to_eval(vars, "update_locked_judgment.xqy")
 
-    def save_judgment_xml(
-        self,
-        judgment_uri: DocumentURIString,
-        judgment_xml: Element,
-        annotation: Optional[str] = None,
-    ) -> requests.Response:
-        """update_judgment uses dls:document-checkout-update-checkin as a single operation"""
-        xml = ElementTree.tostring(judgment_xml)
-
-        uri = self._format_uri_for_marklogic(judgment_uri)
-        vars: query_dicts.UpdateJudgmentDict = {
-            "uri": uri,
-            "judgment": xml.decode("utf-8"),
-            "annotation": annotation or "edited by save_judgment_xml",
-        }
-
-        return self._send_to_eval(vars, "update_judgment.xqy")
-
     def insert_document_xml(
-        self, document_uri: DocumentURIString, document_xml: Element
+        self,
+        document_uri: DocumentURIString,
+        document_xml: Element,
+        annotation: VersionAnnotation,
     ) -> requests.Response:
+        """
+        Insert a new XML document into MarkLogic.
+
+        :param document_uri: The URI to insert the document at
+        :param document_xml: The XML of the document to insert
+        :param annotation: Annotations to record alongside this version
+
+        :return: The response object from MarkLogic
+        """
         xml = ElementTree.tostring(document_xml)
 
         uri = self._format_uri_for_marklogic(document_uri)
+
+        annotation.set_calling_function("insert_document_xml")
+
         vars: query_dicts.InsertDocumentDict = {
             "uri": uri,
             "document": xml.decode("utf-8"),
+            "annotation": annotation.as_json,
         }
 
         return self._send_to_eval(vars, "insert_document.xqy")
+
+    def update_document_xml(
+        self,
+        document_uri: DocumentURIString,
+        document_xml: Element,
+        annotation: VersionAnnotation,
+    ) -> requests.Response:
+        """
+        Updates an existing XML document in MarkLogic with a new version.
+
+        This uses `dls:document-checkout-update-checkin` to perform this in a single operation.
+
+        :param document_uri: The URI of the document to update
+        :param document_xml: The new XML content of the document
+        :param annotation: Annotations to record alongside this version
+
+        :return: The response object from MarkLogic
+        """
+        xml = ElementTree.tostring(document_xml)
+
+        uri = self._format_uri_for_marklogic(document_uri)
+
+        annotation.set_calling_function("update_document_xml")
+
+        vars: query_dicts.UpdateDocumentDict = {
+            "uri": uri,
+            "judgment": xml.decode("utf-8"),
+            "annotation": annotation.as_json,
+        }
+
+        return self._send_to_eval(vars, "update_document.xqy")
 
     def list_judgment_versions(
         self, judgment_uri: DocumentURIString